Address small-player lane review findings
