+++ .pm/tasks/task_96b6823495f44ef39c80f3c8b1a74421.execution.md
@@ -61,3 +61,12 @@ Example:
 - Expected Result: Review package and slice ledger are available for subagent review.
 - Actual Result: Review package generated at .pm/scratch/task_96b6823495f44ef39c80f3c8b1a74421/review-packages/review-b5b0fee23..ce3cc6537.diff; slice ledger path printed.
 - Blocker / Next Action: No blocker; dispatch local role review.
+
+## 2026-06-23 10:37:51 CST / tpm
+- 完成内容: Integrated pre-PR role review findings and completed full required local gate. Added positive regression coverage for grind-only lane truth and forced major-power dependency; split gameplay snapshot helpers/lane derivation out of the oversized runtime snapshot module; fixed a required-gate flaky perf test; updated Cargo.lock to remove a cargo-deny advisory.
+- 遗留事项: Regenerate review package after the follow-up fix commit, run focused re-review for roles with prior actionable findings, then continue prepare-task-pr/PR flow.
+- Action: Review finding closeout. runtime_engineer positive grind_only_flag coverage finding addressed by runtime_gameplay_snapshot_flags_grind_only_after_repeating_same_loop_without_new_leverage. qa_engineer positive forced-dependency coverage finding addressed by runtime_gameplay_snapshot_marks_forced_major_power_dependency_when_no_local_recovery_path_exists, asserting major_power_dependency_status=forced, requires_major_power_sponsorship=yes, recovery_path_kind=none, and repair/rebuild/pivot unavailable. repository_health_engineer stale/polluted package finding addressed by rebase onto refs/remotes/origin/main before this fix pass; the old review package is obsolete and must be regenerated. Required-gate blockers found during verification and fixed: gameplay_snapshot.rs file-size violation fixed by extracting gameplay_snapshot_helpers.rs and gameplay_snapshot_lane.rs; RUSTSEC-2026-0185 for quinn-proto 0.11.14 fixed by Cargo.lock update to quinn-proto 0.11.15; missing viewer node_modules fixed by npm ci before rerunning the Node contract; simulator perf snapshot flaky test fixed by using deterministic LlmLatencyTraceAgent latency instead of a potentially zero-duration WaitingAgent decision.
+- Validation Command: env -u RUSTC_WRAPPER cargo test -p oasis7 runtime_gameplay_snapshot -- --nocapture; env -u RUSTC_WRAPPER cargo test -p oasis7 snapshot_player_gameplay_execution_state_backfills_from_legacy_fields -- --nocapture; env -u RUSTC_WRAPPER cargo test -p oasis7 runtime_gameplay_action_unlocks_first_expansion_tradeoff_after_scale_out -- --nocapture; env -u RUSTC_WRAPPER cargo test -p oasis7 simulator::tests::runner::runner_runtime_perf_snapshot_tracks_tick_and_decision_samples --features test_tier_required -- --nocapture; ./scripts/pm/workflow-lint.sh --task-uid task_96b6823495f44ef39c80f3c8b1a74421 --phase current; git diff --check; ./scripts/ci-tests.sh required
+- Expected Result: Focused runtime/player-gameplay regressions pass, deterministic perf test passes, workflow/diff checks pass, and full required local gate exits 0.
+- Actual Result: Focused runtime_gameplay_snapshot tests passed including the new grind-only and forced-dependency positive cases; legacy snapshot backfill test passed; scale-out unlock test passed; deterministic perf snapshot test passed; workflow-lint passed; git diff --check passed; ./scripts/ci-tests.sh required exited 0 after rerun. The required run emitted existing clippy warnings outside denied correctness/suspicious classes and skipped oasis7_node/oasis7_net clippy as outside the local required baseline.
+- Blocker / Next Action: No current verification blocker. Regenerate review package and collect focused re-review before PR creation.
